slam-rs: make VIO pipeline tests self-contained
