+++ packages/slam-rs/crates/slam-rs/tests/common/mod.rs
@@ -1,35 +1,42 @@
-//! What more than one integration test needs.
-//!
-//! Integration tests are separate crates, so a `tests/common/mod.rs` declared
-//! with `mod common;` is the only way to share code between them. Each test
-//! binary compiles its own copy and uses part of it, which is why the module
-//! allows dead code: the alternative is a `cfg` per item per test. The
-//! `unwrap`/`expect` allows are the module's own rather than each including
-//! binary's, because a fixture that does not parse is a broken checkout and
-//! panicking on it is the report.
+//! Shared integration-test inputs and mathematical references.
+//! Each binary uses a subset; item-level dead-code allowances name a consumer.
 
-#![allow(dead_code, clippy::expect_used, clippy::unwrap_used)]
+#![allow(clippy::expect_used, clippy::unwrap_used)]
 
 use std::path::{Path, PathBuf};
-use std::sync::{Arc, LazyLock};
+use std::sync::LazyLock;
 
-use nalgebra::{DMatrix, DVector, SMatrix, Vector2, Vector3, Vector6};
+use nalgebra::{DMatrix, DVector, Vector2, Vector3, Vector6};
 use serde::Deserialize;
 use slam_rs::calib::{Calibration, CameraModel, Kb4Params};
 use slam_rs::config::VioConfig;
-use slam_rs::estimator::FlowObservations;
 use slam_rs::frontend::tracker::PointsSoA;
 use slam_rs::image::ImageU16;
-use slam_rs::lie::{LieScalar, Se3};
-use slam_rs::types::KeypointId;
+use slam_rs::lie::Se3;
 
+#[allow(
+    dead_code,
+    reason = "used by camera_jacobians; other binaries compile a subset"
+)]
 const MSDMI: &str = include_str!("../fixtures/msdmi_calib.json");
+#[allow(
+    dead_code,
+    reason = "used by camera_jacobians; other binaries compile a subset"
+)]
 const MSDMG: &str = include_str!("../fixtures/msdmg_calib.json");
+#[allow(
+    dead_code,
+    reason = "used by camera_jacobians; other binaries compile a subset"
+)]
 const ROBOCAP: &str = include_str!("../fixtures/robocap-basalt-calib.json");
 
 // ── the fixture directory and the two files every VIO lane reads ───────────
 
 /// `crates/slam-rs/tests/fixtures`.
+#[allow(
+    dead_code,
+    reason = "used by vio_pipeline; other binaries compile a subset"
+)]
 pub fn fixtures() -> PathBuf {
     Path::new(env!("CARGO_MANIFEST_DIR")).join("tests/fixtures")
 }
@@ -37,13 +44,20 @@ pub fn fixtures() -> PathBuf {
 /// `packages/slam-rs/configs`, the shipped VIO configs.
 ///
 /// The crate reads the package's files rather than a copy of its own: these are
-/// the ones `reference_segments.toml` names and the C++ reference runs loaded,
-/// so a lane that drifted from them would compare against a config nothing ran.
+/// the same files used by the Python entry points.
+#[allow(
+    dead_code,
+    reason = "used by vio_pipeline; other binaries compile a subset"
+)]
 pub fn configs() -> PathBuf {
     Path::new(env!("CARGO_MANIFEST_DIR")).join("../../configs")
 }
 
 /// The MSDMI config, which every VIO lane and the whole-pipeline test share.
+#[allow(
+    dead_code,
+    reason = "used by vio_pipeline; other binaries compile a subset"
+)]
 pub fn config() -> VioConfig {
     VioConfig::from_json_str(&std::fs::read_to_string(configs().join("msdmi_config.json")).unwrap())
         .unwrap()
@@ -51,6 +65,10 @@ pub fn config() -> VioConfig {
 
 /// The MSDMI calibration, always `f64`: the estimator casts it to its own
 /// scalar, so a lane that runs `f32` still reads the file's doubles.
+#[allow(
+    dead_code,
+    reason = "used by vio_pipeline; other binaries compile a subset"
+)]
 pub fn calibration() -> Calibration<f64> {
     Calibration::from_json_str(
         &std::fs::read_to_string(fixtures().join("msdmi_calib.json")).unwrap(),
@@ -67,6 +85,10 @@ pub fn calibration() -> Calibration<f64> {
 /// # Panics
 ///
 /// On a name that is not one of the three.
+#[allow(
+    dead_code,
+    reason = "used by camera_jacobians; other binaries compile a subset"
+)]
 pub fn calibration_text(name: &str) -> &'static str {
     match name {
         "msdmi" => MSDMI,
@@ -76,21 +98,23 @@ pub fn calibration_text(name: &str) -> &'static str {
     }
 }
 
-/// Which precision a whole-clip lane runs.
-///
-/// Two lanes select it and, historically, with two vocabularies:
-/// `SLAM_RS_CLIP_SCALAR=f32|f64` for the port's own frontend and
-/// `SLAM_RS_ORACLE_SCALAR=float|double` for the backend replay, so a reader
-/// running both over one clip had to remember which file wanted which word.
-/// [`Self::from_env`] takes either variable and either vocabulary.
+/// Precision selected by the whole-clip runner.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[allow(
+    dead_code,
+    reason = "used by full_clip; other binaries compile a subset"
+)]
 pub enum ClipScalar {
     F32,
     F64,
 }
 
+#[allow(
+    dead_code,
+    reason = "used by full_clip; other binaries compile a subset"
+)]
 impl ClipScalar {
-    /// `SLAM_RS_CLIP_SCALAR` first, then `SLAM_RS_ORACLE_SCALAR`, then
+    /// `SLAM_RS_CLIP_SCALAR`, falling back to
     /// `default`. `f32` and `float` mean the same thing, as do `f64` and
     /// `double`.
     ///
@@ -98,17 +122,14 @@ impl ClipScalar {
     ///
     /// On a value that is none of the four.
     pub fn from_env(default: Self) -> Self {
-        for name in ["SLAM_RS_CLIP_SCALAR", "SLAM_RS_ORACLE_SCALAR"] {
-            let Ok(value) = std::env::var(name) else {
-                continue;
-            };
-            return match value.as_str() {
-                "f32" | "float" => Self::F32,
-                "f64" | "double" => Self::F64,
-                other => panic!("{name} is f32/float or f64/double, not {other}"),
-            };
+        let Ok(value) = std::env::var("SLAM_RS_CLIP_SCALAR") else {
+            return default;
+        };
+        match value.as_str() {
+            "f32" | "float" => Self::F32,
+            "f64" | "double" => Self::F64,
+            other => panic!("SLAM_RS_CLIP_SCALAR is f32/float or f64/double, not {other}"),
         }
-        default
     }
 
     /// `"f32"` or `"f64"`, which is what a written CSV's name carries.
@@ -118,22 +139,14 @@ impl ClipScalar {
             Self::F64 => "f64",
         }
     }
-
-    /// `"float"` or `"double"`, which is how the C++ dump keys its runs.
-    pub fn cpp_name(self) -> &'static str {
-        match self {
-            Self::F32 => "float",
-            Self::F64 => "double",
-        }
-    }
 }
 
-/// What `tests/tools/dump_clip.py` writes beside the pixels.
-///
-/// Both whole-clip lanes read this file — one through its own frontend, one
-/// replaying the C++'s flow stream — so the shape and the dataset-to-config
-/// table live here rather than once typed and once as a `serde_json::Value`.
+/// Clip metadata written beside the input pixels by `tests/tools/dump_clip.py`.
 #[derive(Debug, Deserialize)]
+#[allow(
+    dead_code,
+    reason = "used by full_clip; other binaries compile a subset"
+)]
 pub struct Clip {
     pub segment_id: String,
     pub dataset_name: String,
@@ -148,6 +161,10 @@ pub struct Clip {
     pub imu_samples: usize,
 }
 
+#[allow(
+    dead_code,
+    reason = "used by full_clip; other binaries compile a subset"
+)]
 impl Clip {
     /// `clip.json` from a directory `dump_clip.py` wrote.
     ///
@@ -172,6 +189,10 @@ impl Clip {
 /// # Panics
 ///
 /// On a dataset with no pinned config.
+#[allow(
+    dead_code,
+    reason = "used by full_clip; other binaries compile a subset"
+)]
 pub fn config_for(dataset_name: &str) -> VioConfig {
     if let Some(path) = std::env::var_os("SLAM_RS_CLIP_CONFIG") {
         return VioConfig::from_json_str(&std::fs::read_to_string(path).unwrap()).unwrap();
@@ -184,49 +205,15 @@ pub fn config_for(dataset_name: &str) -> VioConfig {
     VioConfig::from_json_str(&std::fs::read_to_string(configs().join(file)).unwrap()).unwrap()
 }
 
-/// A fixed-size matrix flattened **row major**, which is how every C++ dump
-/// prints one.
-pub fn row_major<const R: usize, const C: usize, S: LieScalar>(m: &SMatrix<S, R, C>) -> Vec<f64> {
-    let mut out: Vec<f64> = Vec::with_capacity(R * C);
-    for r in 0..R {
-        for c in 0..C {
-            out.push(m[(r, c)].to_f64());
-        }
-    }
-    out
-}
-
-/// `‖·‖_F` over any coefficient sequence, which on a vector is `‖·‖`.
-pub fn frobenius<S: LieScalar>(values: impl Iterator<Item = S>) -> f64 {
-    values.map(|v| v.to_f64() * v.to_f64()).sum::<f64>().sqrt()
-}
-
-/// The C++ frontend's keypoints for one frameset, as the estimator takes them.
-///
-/// Both lanes that replay `OracleFlow` need exactly this, and an id the
-/// insertion order would collide on cannot happen: the dump's ids are unique
-/// per camera.
-pub fn observations(flow: &OracleFlow) -> Arc<FlowObservations> {
-    let mut out: FlowObservations = FlowObservations::new(flow.t_ns, flow.cameras.len());
-    for (camera, points) in flow.cameras.iter().enumerate() {
-        let Some(slot) = out.cameras.get_mut(camera) else {
-            continue;
-        };
-        for point in points {
-            slot.insert(KeypointId(point.id), Vector2::new(point.x, point.y));
-        }
-    }
-    Arc::new(out)
-}
-
 // ── the PGM framesets ─────────────────────────────────────────────
 
 /// One camera's image as it sits in a PGM: the raw 8-bit raster and its shape.
 ///
-/// The raster, not an `ImageU16`: `flow_parity` widens it the way basalt's
-/// camera source does, while `vio_parity` hands the bytes straight to
-/// [`slam_rs::ImageView`], and one of those wrapping the other is the only
-/// difference between them.
+/// Used as bytes by VIO and widened to u16 by frontend tests.
+#[allow(
+    dead_code,
+    reason = "used by vio_pipeline; other binaries compile a subset"
+)]
 pub struct Pgm {
     pub width: usize,
     pub height: usize,
@@ -235,6 +222,10 @@ pub struct Pgm {
 
 /// `frame_<NNN>_cam<C>.pgm` under `directory`, in `tools/dump_flow.cpp`'s
 /// layout.
+#[allow(
+    dead_code,
+    reason = "used by vio_pipeline; other binaries compile a subset"
+)]
 pub fn read_pgm(directory: &Path, frame: usize, camera: usize) -> Pgm {
     let path: PathBuf = directory.join(format!("frame_{frame:03}_cam{camera}.pgm"));
     let bytes: Vec<u8> = std::fs::read(&path)
@@ -275,6 +266,10 @@ pub fn read_pgm(directory: &Path, frame: usize, camera: usize) -> Pgm {
 /// framesets, so the fixture location and the widening live here rather than
 /// once per binary. They stay separate binaries: D72 wants the shared CubeCL
 /// pool isolated per test process.
+#[allow(
+    dead_code,
+    reason = "used by gpu_detect; other binaries compile a subset"
+)]
 pub fn mio10_frame(frame: usize, camera: usize) -> ImageU16 {
     let pgm: Pgm = read_pgm(&fixtures().join("flow/frames"), frame, camera);
     let mut image: ImageU16 = ImageU16::zeros(pgm.width, pgm.height).unwrap();
@@ -288,6 +283,10 @@ pub fn mio10_frame(frame: usize, camera: usize) -> ImageU16 {
 
 /// How many consecutive framesets `directory` covers, up to `limit`: a
 /// frameset counts only when every camera's PGM is there.
+#[allow(
+    dead_code,
+    reason = "used by vio_pipeline; other binaries compile a subset"
+)]
 pub fn available_framesets(directory: &Path, cameras: usize, limit: usize) -> usize {
     (0..limit)
         .take_while(|frame| {
@@ -300,130 +299,11 @@ pub fn available_framesets(directory: &Path, cameras: usize, limit: usize) -> us
         .count()
 }
 
-// ── the VIO oracle fixture ─────────────────────────────────────
-
-/// `tools/vio_oracle.cpp`'s dump: one run per precision, plus the C++
-/// frontend's keypoints per frameset. Both VIO lanes read it, so the shape
-/// lives here even though `vio_parity` reads only part of it.
 #[derive(Debug, Deserialize)]
-pub struct Oracle {
-    pub runs: Vec<OracleRun>,
-    pub flow: Vec<OracleFlow>,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct OracleRun {
-    pub scalar: String,
-    pub frames: Vec<OracleFrame>,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct OracleFlow {
-    pub t_ns: i64,
-    pub cameras: Vec<Vec<OraclePoint>>,
-}
-
-/// One tracked keypoint of the C++ frontend's `OpticalFlowResult`.
-///
-/// The fixture also carries the warp's four `linear` coefficients so a reader
-/// can see the whole `AffineCompact2f`; the estimator reads only the
-/// translation, so they are not deserialized.
-#[derive(Debug, Deserialize)]
-pub struct OraclePoint {
-    pub id: u64,
-    pub x: f32,
-    pub y: f32,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct OracleFrame {
-    pub frame: usize,
-    pub t_ns: i64,
-    pub states: Vec<OracleState>,
-    pub poses: Vec<OraclePose>,
-    pub kf_ids: Vec<i64>,
-    pub ltkfs: Vec<i64>,
-    pub num_points_kf: Vec<(i64, i64)>,
-    pub last_state_t_ns: i64,
-    pub frames_after_kf: i32,
-    pub opt_started: bool,
-    pub num_landmarks: usize,
-    pub num_observations: usize,
-    pub num_imu_meas: usize,
-    pub marg_order: Vec<(i64, usize, usize)>,
-    pub marg_digest: OracleDigest,
-    /// The prior itself, row by row, on the framesets `tools/vio_oracle.cpp`
-    /// dumps it: every tenth, so frames 0, 10, 20, 30, 40 and 50 of each lane.
-    #[serde(default)]
-    pub marg_h: Option<Vec<Vec<f64>>>,
-    /// The prior's right-hand side on those same framesets.
-    #[serde(default)]
-    pub marg_b: Option<Vec<f64>>,
-    pub marg: Option<OracleMarg>,
-    pub lm: Vec<OracleLm>,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct OracleState {
-    pub t_ns: i64,
-    pub q: [f64; 4],
-    pub t: [f64; 3],
-    pub vel: [f64; 3],
-    pub bg: [f64; 3],
-    pub ba: [f64; 3],
-    pub linearized: bool,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct OraclePose {
-    pub t_ns: i64,
-    pub q: [f64; 4],
-    pub t: [f64; 3],
-    pub linearized: bool,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct OracleDigest {
-    pub rows: usize,
-    pub cols: usize,
-    pub h_frobenius: f64,
-    pub b_norm: f64,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct OracleMarg {
-    pub states_to_remove: usize,
-    pub last_state_to_marg: i64,
-    pub poses_to_marg: Vec<i64>,
-    pub states_to_marg_all: Vec<i64>,
-    pub states_to_marg_vel_bias: Vec<i64>,
-    pub kfs_to_marg: Vec<i64>,
-    pub idx_to_keep: usize,
-    pub idx_to_marg: usize,
-}
-
-#[derive(Debug, Deserialize)]
-pub struct OracleLm {
-    pub it: i32,
-    pub backtrack: i32,
-    pub error_before: f64,
-    pub error_after: f64,
-    pub vision_error: f64,
-    pub imu_error: f64,
-    pub bg_error: f64,
-    pub ba_error: f64,
-    pub marg_prior_error: f64,
-    pub l_diff: f64,
-    pub f_diff: f64,
-    pub lambda: f64,
-    pub step_norminf: f64,
-    pub solve_attempts: u32,
-    pub step_is_valid: bool,
-    pub step_is_successful: bool,
-}
-
-/// One uncalibrated IMU sample of `vio/imu.json`, as the fixture writes it.
-#[derive(Debug, Deserialize)]
+#[allow(
+    dead_code,
+    reason = "used by vio_pipeline; other binaries compile a subset"
+)]
 pub struct ImuRow {
     pub t_ns: i64,
     pub gyro: [f64; 3],
@@ -431,20 +311,19 @@ pub struct ImuRow {
 }
 
 #[derive(Debug, Deserialize)]
+#[allow(
+    dead_code,
+    reason = "used by vio_pipeline; other binaries compile a subset"
+)]
 struct ImuFixture {
     imu: Vec<ImuRow>,
 }
 
-/// The 1.55 MB oracle, parsed once per test binary: five lanes read it and
-/// `serde_json` is otherwise the slowest thing in them.
-pub static ORACLE: LazyLock<Oracle> = LazyLock::new(|| {
-    let path: PathBuf = fixtures().join("vio/vio_oracle.json");
-    let text: String = std::fs::read_to_string(&path)
-        .unwrap_or_else(|error| panic!("cannot read {}: {error}", path.display()));
-    serde_json::from_str(&text).expect("vio_oracle.json does not match the expected shape")
-});
-
 /// The 1,077 uncalibrated samples of the window, in capture order.
+#[allow(
+    dead_code,
+    reason = "used by vio_pipeline; other binaries compile a subset"
+)]
 pub static IMU: LazyLock<Vec<ImuRow>> = LazyLock::new(|| {
     let path: PathBuf = fixtures().join("vio/imu.json");
     let text: String = std::fs::read_to_string(&path)
@@ -454,18 +333,13 @@ pub static IMU: LazyLock<Vec<ImuRow>> = LazyLock::new(|| {
     fixture.imu
 });
 
-/// The run of one precision, `"double"` or `"float"`.
-pub fn run_named<'a>(oracle: &'a Oracle, scalar: &str) -> &'a OracleRun {
-    oracle
-        .runs
-        .iter()
-        .find(|run| run.scalar == scalar)
-        .unwrap_or_else(|| panic!("the fixture has no {scalar} run"))
-}
-
 /// `KannalaBrandtCamera4<Scalar>::getTestProjections()[0]`
 /// (`basalt-headers/include/basalt/camera/kannala_brandt_camera4.hpp:487-495`),
 /// which is what `test_linearization.cpp:19` puts in both camera slots.
+#[allow(
+    dead_code,
+    reason = "used by linearize_reference; other binaries compile a subset"
+)]
 pub const KB4_TEST_PROJECTION: [f64; 8] = [
     379.045,
     379.008,
@@ -481,8 +355,16 @@ pub const KB4_TEST_PROJECTION: [f64; 8] = [
 ///
 /// A Rust test that flakes is worse than one that is merely differently
 /// arbitrary, so nothing here draws from the system generator.
+#[allow(
+    dead_code,
+    reason = "used by linearize_reference; other binaries compile a subset"
+)]
 pub struct Rng(u64);
 
+#[allow(
+    dead_code,
+    reason = "used by linearize_reference; other binaries compile a subset"
+)]
 impl Rng {
     pub fn new(seed: u64) -> Self {
         Self(seed | 1)
@@ -517,6 +399,10 @@ impl Rng {
 ///
 /// The file supplies only the fields the estimator never touches; everything
 /// read is overwritten here.
+#[allow(
+    dead_code,
+    reason = "used by linearize_reference; other binaries compile a subset"
+)]
 pub fn test_calibration(rng: &mut Rng) -> Calibration<f64> {
     let mut calib: Calibration<f64> = Calibration::from_json_str(MSDMI).unwrap();
     calib.t_i_c = (0..2)
@@ -546,6 +432,10 @@ pub fn test_calibration(rng: &mut Rng) -> Calibration<f64> {
 /// `marginalizeHelperSqrtToSqrt` never forms `JᵀJ`, so squaring its output and
 /// comparing with this is the same argument `VoMargSqrtLinearizationTest` makes
 /// about the linearization (`test_linearization.cpp:379-388`), one level up.
+#[allow(
+    dead_code,
+    reason = "used by marg_window; other binaries compile a subset"
+)]
 pub fn dense_schur(
     h: &DMatrix<f64>,
     b: &DVector<f64>,
@@ -565,84 +455,6 @@ pub fn dense_schur(
     (h_kk - &cross * h_mk, b_k - &cross * b_m)
 }
 
-/// Relative comparison against a C++ dump, tracking the worst case seen.
-///
-/// **Scale.** Every coefficient of an array is compared against the *array's*
-/// largest magnitude, not against itself. That is not laziness: after a
-/// Householder reflection the sub-diagonal entries of the landmark columns are
-/// zero in exact arithmetic and pure cancellation in floating point, so in `f32`
-/// C++ leaves `4.3e-5` where the port leaves `6.5e-3` — both of them noise on a
-/// block whose live coefficients are in the hundreds. basalt's own tests compare
-/// `(H_a - H_b).norm()` for the same reason (`test_linearization.cpp:148-157`).
-pub struct Compare {
-    tolerance: f64,
-    pub worst: f64,
-    pub worst_what: String,
-}
-
-impl Compare {
-    pub fn new(tolerance: f64) -> Self {
-        Self {
-            tolerance,
-            worst: 0.0,
-            worst_what: String::from("(nothing compared)"),
-        }
-    }
-
-    /// One coefficient against a scale the caller chose.
-    pub fn close_scaled(&mut self, got: f64, want: f64, scale: f64, what: &str) {
-        let scale: f64 = scale.max(1.0);
-        let relative: f64 = (got - want).abs() / scale;
-        if relative > self.worst {
-            self.worst = relative;
-            self.worst_what = format!("{what}: got {got:.9e}, want {want:.9e}");
-        }
-        assert!(
-            relative <= self.tolerance,
-            "{what}: got {got:.17e}, want {want:.17e}, relative {relative:.3e} > {:.1e}",
-            self.tolerance
-        );
-    }
-
-    /// One scalar, against its own magnitude.
-    pub fn close(&mut self, got: f64, want: f64, what: &str) {
-        self.close_scaled(got, want, want.abs(), what);
-    }
-
-    pub fn close_slice<S: LieScalar>(&mut self, got: &[S], want: &[f64], what: &str) {
-        assert_eq!(got.len(), want.len(), "{what}: length");
-        let scale: f64 = want.iter().fold(0.0f64, |m, v| m.max(v.abs()));
-        for (i, (g, w)) in got.iter().zip(want.iter()).enumerate() {
-            self.close_scaled(g.to_f64(), *w, scale, &format!("{what}[{i}]"));
-        }
-    }
-
-    /// A row-major matrix from the fixture against a column-major `DMatrix`.
-    pub fn close_matrix<S: LieScalar>(
-        &mut self,
-        got: &DMatrix<S>,
-        want: &[f64],
-        rows: usize,
-        cols: usize,
-        what: &str,
-    ) {
-        assert_eq!(got.nrows(), rows, "{what}: rows");
-        assert_eq!(got.ncols(), cols, "{what}: cols");
-        assert_eq!(want.len(), rows * cols, "{what}: fixture size");
-        let scale: f64 = want.iter().fold(0.0f64, |m, v| m.max(v.abs()));
-        for r in 0..rows {
-            for col in 0..cols {
-                self.close_scaled(
-                    got[(r, col)].to_f64(),
-                    want[r * cols + col],
-                    scale,
-                    &format!("{what}[{r},{col}]"),
-                );
-            }
-        }
-    }
-}
-
 // ── synthetic images and patch positions ───────────────────────────────────
 //
 // Two fields, shared because the tracker's unit tests, the GPU tolerance tests
@@ -653,6 +465,10 @@ impl Compare {
 /// Twelve plane waves between 16 and 56 pixels, in fixed pseudo-random
 /// directions and phases: band-limited, so a shift really does survive down a
 /// pyramid and every patch's `H_se2` is well conditioned.
+#[allow(
+    dead_code,
+    reason = "used by gpu_kernels; other binaries compile a subset"
+)]
 pub fn texture(x: f64, y: f64) -> f64 {
     const WAVES: [(f64, f64, f64); 12] = [
         (16.0, 0.031, 0.11),
@@ -678,6 +494,10 @@ pub fn texture(x: f64, y: f64) -> f64 {
 }
 
 /// [`texture`] rendered into a `u16` image, shifted by `(dx, dy)`.
+#[allow(
+    dead_code,
+    reason = "used by gpu_kernels; other binaries compile a subset"
+)]
 pub fn textured_image(width: usize, height: usize, dx: f32, dy: f32) -> ImageU16 {
     let mut image: ImageU16 = ImageU16::zeros(width, height).expect("a valid image geometry");
     for y in 0..height {
@@ -694,6 +514,10 @@ pub fn textured_image(width: usize, height: usize, dx: f32, dy: f32) -> ImageU16
 /// smooth plane-wave texture above gives almost no corners.
 ///
 /// One LCG plus a `sin`/`cos` wave, so it is the same field on every machine.
+#[allow(
+    dead_code,
+    reason = "used by fast_model; other binaries compile a subset"
+)]
 pub fn cornered_bytes(width: usize, height: usize) -> Vec<u8> {
     let mut out: Vec<u8> = vec![0u8; width * height];
     let mut state: u32 = 0x1234_5678;
@@ -712,6 +536,10 @@ pub fn cornered_bytes(width: usize, height: usize) -> Vec<u8> {
 ///
 /// The detector reads `pixel >> 8`, so this is the same field the CPU sweep and
 /// the GPU score kernel see.
+#[allow(
+    dead_code,
+    reason = "used by gpu_detect; other binaries compile a subset"
+)]
 pub fn cornered_image(width: usize, height: usize) -> ImageU16 {
     let bytes: Vec<u8> = cornered_bytes(width, height);
     let mut image: ImageU16 = ImageU16::zeros(width, height).expect("a valid image geometry");
@@ -726,6 +554,10 @@ pub fn cornered_image(width: usize, height: usize) -> ImageU16 {
 /// A grid of source positions well inside a `size` x `size` frame, spaced so no
 /// two patches overlap and every one is far enough from the border for the
 /// coarsest level's 52-tap pattern.
+#[allow(
+    dead_code,
+    reason = "used by gpu_kernels; other binaries compile a subset"
+)]
 pub fn grid_positions(size: usize) -> PointsSoA {
     let mut positions: PointsSoA = PointsSoA::with_capacity(256);
     let mut y: usize = 96;
@@ -754,12 +586,24 @@ use slam_rs::pyramid::{CpuPyramidBuilder, PyramidBuilder, PyramidU16};
 use std::collections::BTreeMap;
 
 /// The synthetic rig's frame size, shared by `flow_rig` and `dotted_image`.
+#[allow(
+    dead_code,
+    reason = "used by flow_frontend; other binaries compile a subset"
+)]
 pub const FLOW_WIDTH: usize = 200;
 /// The synthetic rig's frame height.
+#[allow(
+    dead_code,
+    reason = "used by flow_frontend; other binaries compile a subset"
+)]
 pub const FLOW_HEIGHT: usize = 200;
 
 /// `count` identical pinhole cameras 5 cm apart along `x`, all seeing a
 /// `FLOW_WIDTH` x `FLOW_HEIGHT` frame.
+#[allow(
+    dead_code,
+    reason = "used by flow_frontend; other binaries compile a subset"
+)]
 pub fn flow_rig(count: usize) -> Calibration<f64> {
     let intrinsics: CameraModel<f64> = CameraModel::Pinhole(PinholeParams {
         fx: 180.0,
@@ -789,6 +633,10 @@ pub fn flow_rig(count: usize) -> Calibration<f64> {
 /// basalt's shipped configuration, with the matching guess set to the same
 /// pixel so that `flow_rig`'s cameras — which see identical frames — really do
 /// match.
+#[allow(
+    dead_code,
+    reason = "used by flow_frontend; other binaries compile a subset"
+)]
 pub fn flow_config() -> VioConfig {
     VioConfig {
         optical_flow_matching_guess_type: MatchingGuessType::SamePixel,
@@ -799,6 +647,10 @@ pub fn flow_config() -> VioConfig {
 /// Bright 5x5 squares on a regular lattice, the whole frame shifted by `shift`
 /// pixels: four strong FAST corners each, and enough texture in between for the
 /// KLT to follow them.
+#[allow(
+    dead_code,
+    reason = "used by flow_frontend; other binaries compile a subset"
+)]
 pub fn dotted_image(shift: i32) -> ImageU16 {
     let mut image: ImageU16 = ImageU16::zeros(FLOW_WIDTH, FLOW_HEIGHT).expect("a valid geometry");
     for y in 0..FLOW_HEIGHT {
@@ -829,6 +681,10 @@ pub fn dotted_image(shift: i32) -> ImageU16 {
 }
 
 /// A CPU pyramid of `image` with `levels` halvings on top of level 0.
+#[allow(
+    dead_code,
+    reason = "used by klt_tracker; other binaries compile a subset"
+)]
 pub fn pyramid_of(image: &ImageU16, levels: usize) -> PyramidU16 {
     let mut pyramid: PyramidU16 =
         PyramidU16::with_capacity(image.width(), image.height(), levels).expect("a valid geometry");
