+++ packages/slam-rs/crates/slam-rs/tests/vio_pipeline.rs
@@ -1,72 +1,27 @@
-//! The V3 gate: the whole VIO, Rust frontend and Rust estimator, against
-//! basalt's per-frame poses.
-//!
-//! `vio_oracle.rs` isolates the backend by replaying the C++'s own keypoints.
-//! This file does not: it runs [`Vio::track`] on the PGMs and lets the port's
-//! frontend choose the keypoints, which is what a caller gets. The two
-//! frontends select different corners inside a grid cell that offers more than
-//! the budget, because basalt detects with `cv::FAST` and the port with
-//! kornia-rs (D09, D45), so 3 to 4.5 % of the keypoints differ — and a keypoint
-//! set that differs by one id changes which landmarks get triangulated, which
-//! changes the window. So this is a **tolerance** gate on the trajectory, not a
-//! parity gate on the numbers, and the tolerances below are the measurement
-//! plus a margin.
-//!
-//! ## What it needs, and why it skips
-//!
-//! The 60 framesets are 107 MB of PGM and are not committed; point
-//! `SLAM_RS_VIO_FRAMES_DIR` at a directory holding
-//! `frame_<NNN>_cam<C>.pgm` (`tools/dump_flow.cpp`'s layout, which
-//! `slam_rs.catalog_feed` writes on the frozen gray8 decode path) and the test
-//! runs, and it has to hold **all** 60: a prefix is refused rather than
-//! measured, because a short run cannot show a late divergence. Without the
-//! variable the test prints why and passes — the three framesets
-//! `flow/frames/` does carry are too few to reach `opt_started`, which needs
-//! five states.
-//!
-//! Timestamps and the IMU window come from the committed fixtures — the
-//! oracle's `flow[].t_ns` and `vio/imu.json` — so the directory only has to
-//! hold pixels.
-//!
-//! ## What is measured
-//!
-//! Per frameset, against the C++ `float` run's newest state: the position
-//! difference in metres and the rotation difference in degrees, both absolute,
-//! plus the distance travelled so the numbers can be read in context. The
-//! estimator runs `f32`, the precision basalt ships (Q07). On the smoke segment
-//! all 60 framesets reach `Tracking` and the worst differences are 6.14e-4 m
-//! and 0.219° over 1.49 cm of travel — the segment is a panorama, so the
-//! rotation is the number with meaning.
+//! Self-contained whole-pipeline, retry, and prior-comparison tests.
 
 #![allow(clippy::unwrap_used, clippy::expect_used)]
 
 use std::path::PathBuf;
 
-use nalgebra::Vector3;
+use nalgebra::{DMatrix, DVector, SymmetricEigen};
 
 use slam_rs::config::VioConfig;
+use slam_rs::estimator::FrameUpdateOutcome;
 use slam_rs::frontend::flow::{FrontendError, FrontendOptions};
-use slam_rs::lie::So3;
+use slam_rs::lie::LieScalar;
 use slam_rs::{Backend, ImageView, Vio, VioError, VioResult, VioStatus};
 
 mod common;
-use common::{IMU, ORACLE, OracleFrame, OracleState, Pgm, run_named};
+use common::{IMU, Pgm};
+use std::sync::LazyLock;
 
-/// How far the port's rig position may sit from basalt's, in metres.
-///
-/// Measured over the 60 framesets: worst 6.14e-4 m. The segment is a panorama —
-/// 1.49 cm of translation over 1.09 s — so this is not a drift ratio, it is how
-/// far two nearly-stationary trajectories part; five times the measurement.
-const POSITION_TOLERANCE_M: f64 = 3e-3;
-
-/// How far the port's rig orientation may sit from basalt's, in degrees.
-///
-/// Measured worst: 0.219°, and on a panorama this is the number that matters.
-/// It is two orders above the backend-only gate's 8.4e-5 relative rotation
-/// because the two frontends pick different corners (D45) and a different
-/// keypoint set gives a different set of triangulated landmarks; four times
-/// the measurement.
-const ROTATION_TOLERANCE_DEG: f64 = 1.0;
+static TIMESTAMPS: LazyLock<Vec<i64>> = LazyLock::new(|| {
+    include_str!("fixtures/flow/frames/timestamps.txt")
+        .lines()
+        .map(|line| line.parse().unwrap())
+        .collect()
+});
 
 /// Framesets `tests/fixtures/flow/frames/` carries.
 const COMMITTED_FRAMESETS: usize = 3;
@@ -74,7 +29,7 @@ const COMMITTED_FRAMESETS: usize = 3;
 /// Side of the square frame the wrong-size probe offers; the rig's is 960.
 const CROPPED_SIDE: usize = 64;
 
-/// The last of them, `ORACLE.flow[2].t_ns`.
+/// The last of them, `TIMESTAMPS[2]`.
 const LAST_COMMITTED_T_NS: i64 = 37_012_000;
 
 /// How far the IMU is pushed for the committed run: one frameset interval past
@@ -88,7 +43,7 @@ const COMMITTED_IMU_HORIZON_NS: i64 = 40_000_000;
 
 /// The pipeline both tests drive: the fixture's config and calibration, one
 /// frontend thread so the reduction shape is fixed (D31).
-fn pipeline() -> Vio<f32> {
+fn pipeline<S: LieScalar>() -> Vio<S> {
     Vio::new(
         common::config(),
         common::calibration(),
@@ -150,9 +105,8 @@ fn view(pgm: &Pgm) -> ImageView<'_> {
 /// The whole pipeline needs real pixels: `Vio::track` runs the KLT before the
 /// estimator sees anything, and a constant frame detects no corners. Three is
 /// what `tests/fixtures/flow/frames/` carries, so this cannot reach
-/// `opt_started` (five states) — that is what the sixty-frameset gate above
-/// is for.
-fn drive_the_committed_framesets(vio: &mut Vio<f32>) -> Vec<VioResult> {
+/// `opt_started` (five states). The schedule test extends the input by holding the last image.
+fn drive_the_committed_framesets<S: LieScalar>(vio: &mut Vio<S>) -> Vec<VioResult> {
     let directory: PathBuf = common::fixtures().join("flow/frames");
     let cameras: usize = common::calibration().t_i_c.len();
     let framesets: usize = common::available_framesets(&directory, cameras, COMMITTED_FRAMESETS);
@@ -164,8 +118,7 @@ fn drive_the_committed_framesets(vio: &mut Vio<f32>) -> Vec<VioResult> {
     {
         vio.push_imu(row.t_ns, row.gyro, row.accel).unwrap();
     }
-    ORACLE
-        .flow
+    TIMESTAMPS
         .iter()
         .take(framesets)
         .enumerate()
@@ -174,21 +127,25 @@ fn drive_the_committed_framesets(vio: &mut Vio<f32>) -> Vec<VioResult> {
                 .map(|camera| common::read_pgm(&directory, frame, camera))
                 .collect();
             let views: Vec<ImageView<'_>> = rasters.iter().map(view).collect();
-            vio.track(flow.t_ns, &views).unwrap()
+            vio.track(*flow, &views).unwrap()
         })
         .collect()
 }
 
 /// The frontend and the estimator are wired together, and a repeat run is
-/// bit-identical (D17) — `vio_oracle.rs`'s determinism gate replays the C++'s
-/// keypoints, so it cannot see the frontend.
+/// bit-identical (D17).
 ///
 /// In the default lane: three framesets of 960x960 KLT twice over is 0.31 s
 /// with the optimized test profile (`Cargo.toml`), where it was 14 s of the
 /// suite's 15 s budget unoptimized, which is why it used to be `#[ignore]`d.
 #[test]
 fn the_whole_pipeline_tracks_and_repeats_bit_identically() {
-    let mut vio: Vio<f32> = pipeline();
+    check_pipeline::<f32>();
+    check_pipeline::<f64>();
+}
+
+fn check_pipeline<S: LieScalar>() {
+    let mut vio: Vio<S> = pipeline();
     let results: Vec<VioResult> = drive_the_committed_framesets(&mut vio);
 
     assert_eq!(
@@ -197,10 +154,7 @@ fn the_whole_pipeline_tracks_and_repeats_bit_identically() {
     );
     assert_eq!(
         results.iter().map(|r| r.t_ns).collect::<Vec<i64>>(),
-        ORACLE.flow[..COMMITTED_FRAMESETS]
-            .iter()
-            .map(|flow| flow.t_ns)
-            .collect::<Vec<i64>>()
+        TIMESTAMPS[..COMMITTED_FRAMESETS].to_vec()
     );
     let estimator = vio.estimator();
     assert_eq!(estimator.snapshot().states.len(), COMMITTED_FRAMESETS);
@@ -215,7 +169,7 @@ fn the_whole_pipeline_tracks_and_repeats_bit_identically() {
     assert!(stats.num_landmarks > 0);
 
     assert_eq!(
-        drive_the_committed_framesets(&mut pipeline()),
+        drive_the_committed_framesets(&mut pipeline::<S>()),
         results,
         "a repeat run is not bit-identical"
     );
@@ -313,7 +267,7 @@ fn a_wrong_size_frameset_is_refused_without_moving_the_pipeline() {
         .enumerate()
         .map(|(frame, raster)| {
             let views: Vec<ImageView<'_>> = raster.iter().map(view).collect();
-            clean.track(ORACLE.flow[frame].t_ns, &views).unwrap()
+            clean.track(TIMESTAMPS[frame], &views).unwrap()
         })
         .collect();
 
@@ -331,7 +285,7 @@ fn a_wrong_size_frameset_is_refused_without_moving_the_pipeline() {
     push_imu_through(&mut probed, 0, COMMITTED_IMU_HORIZON_NS);
     let mut got: Vec<VioResult> = Vec::new();
     for (frame, raster) in rasters.iter().enumerate() {
-        let t_ns: i64 = ORACLE.flow[frame].t_ns;
+        let t_ns: i64 = TIMESTAMPS[frame];
 
         let before: u64 = fingerprint(&probed);
         let refused: VioError = probed.track(t_ns, &wrong).unwrap_err();
@@ -392,7 +346,7 @@ fn a_refused_frameset_is_retried_bit_identically() {
         .enumerate()
         .map(|(frame, raster)| {
             let views: Vec<ImageView<'_>> = raster.iter().map(view).collect();
-            ahead.track(ORACLE.flow[frame].t_ns, &views).unwrap()
+            ahead.track(TIMESTAMPS[frame], &views).unwrap()
         })
         .collect();
 
@@ -402,7 +356,7 @@ fn a_refused_frameset_is_retried_bit_identically() {
     let mut next: usize = 0;
     let mut got: Vec<VioResult> = Vec::new();
     for (frame, raster) in rasters.iter().enumerate() {
-        let t_ns: i64 = ORACLE.flow[frame].t_ns;
+        let t_ns: i64 = TIMESTAMPS[frame];
         let views: Vec<ImageView<'_>> = raster.iter().map(view).collect();
 
         let before: u64 = fingerprint(&behind);
@@ -423,7 +377,7 @@ fn a_refused_frameset_is_retried_bit_identically() {
         // The last horizon is the reference run's, so both runs end holding the
         // same samples.
         let horizon: i64 = if frame + 1 < COMMITTED_FRAMESETS {
-            ORACLE.flow[frame + 1].t_ns
+            TIMESTAMPS[frame + 1]
         } else {
             COMMITTED_IMU_HORIZON_NS
         };
@@ -439,113 +393,144 @@ fn a_refused_frameset_is_retried_bit_identically() {
     );
 }
 
-/// The whole pipeline over **all 60** framesets, reporting the per-frame
-/// difference from basalt's own trajectory.
-///
-/// Every frameset has to reach `Tracking` and every pose is compared: a gate
-/// that skipped the ones that did not would pass on a single pose while the
-/// port lost tracking on the other 59.
-#[test]
-fn the_whole_vio_follows_the_cpp_trajectory() {
-    let Some(directory) = std::env::var_os("SLAM_RS_VIO_FRAMES_DIR").map(PathBuf::from) else {
-        println!(
-            "skipped: set SLAM_RS_VIO_FRAMES_DIR to a directory of frame_<NNN>_cam<C>.pgm; the \
-             three committed framesets cannot reach opt_started, which needs five states"
-        );
-        return;
-    };
+const PRIOR_NULL_RATIO: f64 = 1e-5;
+const PRIOR_TOLERANCE_F64: f64 = 5e-6;
+const PRIOR_TOLERANCE_F32: f64 = 3e-1;
 
-    let run = run_named(&ORACLE, "float");
-    let cameras: usize = common::calibration().t_i_c.len();
-    let available: usize = common::available_framesets(&directory, cameras, ORACLE.flow.len());
-    assert_eq!(
-        available,
-        ORACLE.flow.len(),
-        "{} holds {available} of the {} framesets the fixture describes; a short directory would \
-         make this gate pass on a prefix",
-        directory.display(),
-        ORACLE.flow.len()
-    );
+fn prior_deviation(
+    prior: &DMatrix<f64>,
+    rhs: &DVector<f64>,
+    reference: &DMatrix<f64>,
+    reference_rhs: &DVector<f64>,
+) -> (f64, f64) {
+    let information: DMatrix<f64> = prior.transpose() * prior;
+    let reference_information: DMatrix<f64> = reference.transpose() * reference;
+    let term: DVector<f64> = prior.transpose() * rhs;
+    let reference_term: DVector<f64> = reference.transpose() * reference_rhs;
 
-    let mut vio: Vio<f32> = pipeline();
-    for row in IMU.iter() {
-        vio.push_imu(row.t_ns, row.gyro, row.accel).unwrap();
-    }
+    let eigen: SymmetricEigen<f64, nalgebra::Dyn> =
+        SymmetricEigen::new(reference_information.clone());
+    let floor: f64 =
+        PRIOR_NULL_RATIO * PRIOR_NULL_RATIO * eigen.eigenvalues.iter().copied().fold(0.0, f64::max);
+    let whitener: DMatrix<f64> = &eigen.eigenvectors
+        * DMatrix::from_diagonal(&eigen.eigenvalues.map(|value| 1.0 / value.max(floor).sqrt()));
 
-    let mut worst_position: f64 = 0.0;
-    let mut worst_rotation_deg: f64 = 0.0;
-    let mut travelled: f64 = 0.0;
-    let mut previous: Option<Vector3<f64>> = None;
-    let mut compared: usize = 0;
+    let error: DMatrix<f64> =
+        &(whitener.transpose() * (information - &reference_information)) * &whitener;
+    let deviation: f64 = SymmetricEigen::new(error).eigenvalues.amax();
 
-    for (frame, flow) in ORACLE.flow.iter().take(available).enumerate() {
-        let rasters: Vec<Pgm> = (0..cameras)
-            .map(|camera| common::read_pgm(&directory, frame, camera))
-            .collect();
-        let views: Vec<ImageView<'_>> = rasters.iter().map(view).collect();
-        let result = vio.track(flow.t_ns, &views).unwrap();
-        assert_eq!(result.t_ns, flow.t_ns, "frame {frame}: timestamp");
-        // No frameset may legitimately miss `Tracking` here, and skipping one
-        // would let the tolerances below pass on a single pose. The whole IMU
-        // is pushed above and its last sample (1_091_997_574 ns) is past the
-        // last frameset (1_091_905_000 ns), so every frameset is covered and
-        // none can report `NeedMoreImu`; and the estimator initialises inside
-        // the same `process_frame` that measures (`:263-296`), so frameset 0 is
-        // `Tracking` too and `NotInitialised` is unreachable from `Vio`. The
-        // C++ dump agrees: all 60 of its `frames` carry a state, `frames[0]`
-        // included.
-        assert_eq!(
-            result.status,
-            VioStatus::Tracking,
-            "frame {frame} at {} ns lost tracking",
-            flow.t_ns
-        );
+    let residual: DVector<f64> = whitener.transpose() * (term - &reference_term);
+    let scale: f64 = (whitener.transpose() * reference_term).norm().max(1.0);
+    (deviation, residual.norm() / scale)
+}
 
-        // basalt's newest state at this frameset. `last_state_t_ns` is the
-        // frameset itself once `measure` has run, so this is the same pose the
-        // port reports.
-        let expected: &OracleFrame = &run.frames[frame];
-        let newest: &OracleState = expected
-            .states
-            .iter()
-            .find(|state| state.t_ns == expected.last_state_t_ns)
-            .unwrap_or_else(|| panic!("frame {frame}: the newest state is not in the window"));
-
-        let pose: [f64; 7] = result.world_from_rig;
-        let position: Vector3<f64> = Vector3::new(pose[0], pose[1], pose[2]);
-        if let Some(before) = previous {
-            travelled += (position - before).norm();
+fn reference_prior() -> (DMatrix<f64>, DVector<f64>) {
+    let directions: [[f64; 6]; 4] = [
+        [1.0, 1.0, 0.0, 0.0, 0.0, 0.0],
+        [1.0, -1.0, 0.0, 0.0, 0.0, 0.0],
+        [0.0, 0.0, 1.0, 1.0, 0.0, 0.0],
+        [0.0, 0.0, 1.0, -1.0, 0.0, 0.0],
+    ];
+    let singular: [f64; 4] = [1.0e4, 1.0e2, 1.0e1, 3.0];
+    let mut prior: DMatrix<f64> = DMatrix::zeros(5, 6);
+    for (row, (direction, sigma)) in directions.iter().zip(singular).enumerate() {
+        for (column, value) in direction.iter().enumerate() {
+            prior[(row, column)] = value * sigma * std::f64::consts::FRAC_1_SQRT_2;
         }
-        previous = Some(position);
-
-        let offset: f64 = (position - Vector3::from(newest.t)).norm();
-        // The relative rotation's angle, from Sophus's own log.
-        let mine: So3<f64> = So3::from_quaternion_xyzw(pose[3], pose[4], pose[5], pose[6]).unwrap();
-        let theirs: So3<f64> =
-            So3::from_quaternion_xyzw(newest.q[0], newest.q[1], newest.q[2], newest.q[3]).unwrap();
-        let angle_deg: f64 = (mine.inverse() * theirs).log().norm().to_degrees();
-
-        worst_position = worst_position.max(offset);
-        worst_rotation_deg = worst_rotation_deg.max(angle_deg);
-        compared += 1;
     }
+    (prior, DVector::from_vec(vec![1.0, -2.0, 0.5, 0.25, 0.0]))
+}
 
-    println!(
-        "whole-VIO parity over {compared} of {available} framesets: worst position \
-         {worst_position:.4e} m, worst rotation {worst_rotation_deg:.4e} deg, {travelled:.4} m \
-         travelled"
-    );
-    assert_eq!(
-        compared,
-        ORACLE.flow.len(),
-        "every frameset has to be compared, not just the ones that tracked"
-    );
+#[test]
+fn a_constraint_hidden_beside_a_zero_row_fails_the_prior_comparison() {
+    let (reference, reference_rhs) = reference_prior();
+    let mut candidate: DMatrix<f64> = reference.clone().insert_row(reference.nrows(), 0.0);
+    candidate[(reference.nrows(), 4)] = 0.5;
+    let rhs: DVector<f64> = reference_rhs.clone().insert_row(reference_rhs.nrows(), 0.0);
+
+    // What the shape said about it: one surplus row, and a smallest row norm of
+    // zero, from the row the reference already had.
+    assert_eq!(candidate.nrows(), reference.nrows() + 1);
+    let smallest: f64 = (0..candidate.nrows())
+        .map(|row| candidate.row(row).norm())
+        .fold(f64::INFINITY, f64::min);
+    assert_eq!(smallest, 0.0);
+
+    // The control first: the reference against itself is exactly equal.
+    let (same, _) = prior_deviation(&reference, &reference_rhs, &reference, &reference_rhs);
     assert!(
-        worst_position <= POSITION_TOLERANCE_M,
-        "the trajectory drifted {worst_position:.4e} m from basalt's over {travelled:.4} m"
+        same < PRIOR_TOLERANCE_F64,
+        "the reference moved: {same:.3e}"
     );
+
+    let (deviation, _) = prior_deviation(&candidate, &rhs, &reference, &reference_rhs);
     assert!(
-        worst_rotation_deg <= ROTATION_TOLERANCE_DEG,
-        "the orientation drifted {worst_rotation_deg:.4e} deg from basalt's"
+        deviation > PRIOR_TOLERANCE_F32,
+        "a 0.5 constraint in a direction the reference leaves free scored {deviation:.3e}, \
+         inside the loosest lane tolerance {PRIOR_TOLERANCE_F32:.1e}"
     );
+}
+
+#[test]
+fn a_lost_direction_fails_the_prior_comparison_at_the_same_shape_and_norm() {
+    let (reference, reference_rhs) = reference_prior();
+    let mut candidate: DMatrix<f64> = reference.clone();
+    candidate.row_mut(3).fill(0.0);
+    candidate *= (reference.norm_squared() / candidate.norm_squared()).sqrt();
+
+    // What the shape and the digest say about it: nothing.
+    assert_eq!(candidate.shape(), reference.shape());
+    assert!((candidate.norm() - reference.norm()).abs() <= 1e-9 * reference.norm());
+
+    let (deviation, _) = prior_deviation(&candidate, &reference_rhs, &reference, &reference_rhs);
+    assert!(
+        deviation > PRIOR_TOLERANCE_F32,
+        "dropping the weakest of four constrained directions scored {deviation:.3e}, inside the \
+         loosest lane tolerance {PRIOR_TOLERANCE_F32:.1e}"
+    );
+}
+
+#[test]
+fn the_frame_update_lane_is_deterministic_and_takes_its_branch() {
+    fn drive<S: LieScalar>() -> Vec<(VioResult, FrameUpdateOutcome, bool, usize)> {
+        let mut config = common::config();
+        config.port_frame_update_max_iterations = 5;
+        let mut vio =
+            Vio::<S>::new(config, common::calibration(), FrontendOptions::default()).unwrap();
+        for row in IMU.iter() {
+            vio.push_imu(row.t_ns, row.gyro, row.accel).unwrap();
+        }
+        let directory = common::fixtures().join("flow/frames");
+        let mut output = Vec::new();
+        let mut optimized = false;
+        for (index, &time) in TIMESTAMPS.iter().enumerate() {
+            // Only three raster framesets are shipped. Hold the final image for
+            // the remaining timestamps to exercise scheduling beyond warmup.
+            // This is a schedule probe, not an accuracy measurement.
+            let rasters: Vec<_> = (0..2)
+                .map(|camera| common::read_pgm(&directory, index.min(2), camera))
+                .collect();
+            let views: Vec<_> = rasters.iter().map(view).collect();
+            let result = vio.track(time, &views).unwrap();
+            assert_eq!(result.status, VioStatus::Tracking);
+            let stats = vio.last_stats().unwrap();
+            if !optimized || stats.took_keyframe {
+                assert_eq!(stats.frame_update, FrameUpdateOutcome::NotAttempted);
+            } else {
+                assert_eq!(stats.frame_update, FrameUpdateOutcome::Taken);
+                assert!(stats.lm.len() <= 6);
+            }
+            optimized = stats.opt_started;
+            output.push((
+                result,
+                stats.frame_update,
+                stats.took_keyframe,
+                stats.lm.len(),
+            ));
+        }
+        assert!(output.iter().any(|row| row.1 == FrameUpdateOutcome::Taken));
+        output
+    }
+    assert_eq!(drive::<f32>(), drive::<f32>());
+    assert_eq!(drive::<f64>(), drive::<f64>());
 }
